Moving simulator h files to include
