+++ include/coreir/simulator/simulator.h
@@ -1,3 +1,25 @@
 #pragma once
 
-#include "../../../src/simulator/sim.hpp"
+#include "op_graph.hpp"
+
+namespace CoreIR {
+
+  typedef NGraph ThreadGraph;
+
+  int numThreads(const ThreadGraph& g);
+
+  std::string printCode(const std::deque<vdisc>& topoOrder,
+			NGraph& g,
+			CoreIR::Module* mod,
+			const std::string& baseName);
+
+  std::string printDecl(CoreIR::Module* mod,
+			const NGraph& g);
+
+  std::vector<std::pair<CoreIR::Type*, std::string> >
+  sortedSimArgumentPairs(Module& mod);
+
+
+  ThreadGraph buildThreadGraph(const NGraph& opG);
+
+}
